Provide working implementation and configuration of TypeScript LARA Plugin API
[#165062314]
