+++ lara-plugin-api/src/global.d.ts
@@ -1,0 +1,3 @@
+declare module "sidebar";
+declare module "plugins";
+declare module "text-decorator";
